requirements.txt
python-dotenv
google-generativeai
python-telegram-bot